- feat     : provider 기반 소셜 프로필/닉네임 조회 및 달성률 정수화

+++ apps/users/serializers/profile_serializers.py
@@ -53,7 +53,7 @@ class MeActivitySummaryDaysResponseSerializer(serializers.Serializer[Any]):
 class MeActivitySummaryAchievementRateDetailSerializer(serializers.Serializer[Any]):
     total_goals_count = serializers.IntegerField()
     completed_goals_count = serializers.IntegerField()
-    total_achievement_rate = serializers.FloatField()
+    total_achievement_rate = serializers.IntegerField()
 
 
 class MeActivitySummaryAchievementRateResponseSerializer(serializers.Serializer[Any]):
